add: add Binder interface and refactor StoreViewBinder
