+++ gemini-android/src/main/java/com/haroncode/gemini/android/binder/Binder.kt
@@ -1,0 +1,8 @@
+package com.haroncode.gemini.android.binder
+
+import androidx.lifecycle.LifecycleOwner
+
+interface Binder<View : LifecycleOwner> {
+
+    fun bind(view: View)
+}
