+++ gemini-android/src/main/java/com/haroncode/gemini/android/binder/StoreViewBinding.kt
@@ -1,0 +1,31 @@
+package com.haroncode.gemini.android.binder
+
+import androidx.savedstate.SavedStateRegistryOwner
+import com.haroncode.gemini.android.lifecycle.LifecycleStrategy
+import com.haroncode.gemini.android.lifecycle.StartStopStrategy
+import com.haroncode.gemini.connector.ConnectionRulesFactory
+
+/**
+ * @author HaronCode
+ * @author kdk96
+ */
+object StoreViewBinding {
+
+    @JvmStatic
+    fun <T : SavedStateRegistryOwner> with(
+        lifecycleStrategy: LifecycleStrategy = StartStopStrategy,
+        factoryProvider: () -> ConnectionRulesFactory<T>,
+    ): Binder<T> = SimpleBinder(
+        factoryProvider = factoryProvider,
+        lifecycleStrategy = lifecycleStrategy,
+    )
+
+    @JvmStatic
+    fun <T : SavedStateRegistryOwner> withRestore(
+        lifecycleStrategy: LifecycleStrategy = StartStopStrategy,
+        factoryProvider: () -> ConnectionRulesFactory<T>,
+    ): Binder<T> = RestoreBinder(
+        factoryProvider = factoryProvider,
+        lifecycleStrategy = lifecycleStrategy
+    )
+}
